Atallah & Mohammad O

+++ restaurants/views.py
@@ -3,6 +3,10 @@ from .models import Restaurant, Item
 from .forms import RestaurantForm, ItemForm, SignupForm, SigninForm
 from django.contrib.auth import login, authenticate, logout
 
+
+def noaccess(request):
+    return render(request, 'no-access.html')
+
 def signup(request):
     form = SignupForm()
     if request.method == 'POST':
@@ -59,6 +63,8 @@ def restaurant_detail(request, restaurant_id):
     return render(request, 'detail.html', context)
 
 def restaurant_create(request):
+    if request.user.is_anonymous:
+        return redirect('signin')
     form = RestaurantForm()
     if request.method == "POST":
         form = RestaurantForm(request.POST, request.FILES)
@@ -73,8 +79,14 @@ def restaurant_create(request):
     return render(request, 'create.html', context)
 
 def item_create(request, restaurant_id):
-    form = ItemForm()
+    if request.user.is_anonymous:
+        return redirect('signin')
     restaurant = Restaurant.objects.get(id=restaurant_id)
+    if not (request.user.is_staff or request.user == restaurant.owner):
+        return redirect('no-access')
+
+    form = ItemForm()
+    
     if request.method == "POST":
         form = ItemForm(request.POST)
         if form.is_valid():
@@ -89,20 +101,28 @@ def item_create(request, restaurant_id):
     return render(request, 'item_create.html', context)
 
 def restaurant_update(request, restaurant_id):
-    restaurant_obj = Restaurant.objects.get(id=restaurant_id)
-    form = RestaurantForm(instance=restaurant_obj)
+    if request.user.is_anonymous:
+        return redirect('signin')
+    restaurant = Restaurant.objects.get(id=restaurant_id)
+    if not (request.user.is_staff or request.user == restaurant.owner):
+        return redirect('no-access')
+    form = RestaurantForm(instance=restaurant)
     if request.method == "POST":
-        form = RestaurantForm(request.POST, request.FILES, instance=restaurant_obj)
+        form = RestaurantForm(request.POST, request.FILES, instance=restaurant)
         if form.is_valid():
             form.save()
             return redirect('restaurant-list')
     context = {
-        "restaurant_obj": restaurant_obj,
+        "restaurant_obj": restaurant,
         "form":form,
     }
     return render(request, 'update.html', context)
 
 def restaurant_delete(request, restaurant_id):
+    if request.user.is_anonymous:
+        return redirect('signin')
+    if not request.user.is_staff:
+        return redirect('no-access')
     restaurant_obj = Restaurant.objects.get(id=restaurant_id)
     restaurant_obj.delete()
     return redirect('restaurant-list')
